+++ lib/view/screen/checkout/inner_widget/checkout_top_section.dart
@@ -63,7 +63,7 @@ class _CheckoutTopSectionState extends State<CheckoutTopSection> {
                 activeColor: AppColors.blackPrimary,
               ),
               dotsCount: data?[widget.index].residenceId?.photo?.length ?? 0,
-              position:currentPosition,
+              position:currentPosition.toDouble(),
             )
           ],
         ),
